Backend/requirements.txt
aiofiles
alembic
biopython
celery[redis]
clustalw
fastapi[all]
muscle
passlib[bcrypt]
psycopg2-binary
asyncpg
python-dotenv
python-jose[cryptography]
python-multipart
redis-py
sqlalchemy
uvicorn[standard]
xmltodict
aiohttp